add 2 solutions for day 16, one slightly faster, but more verbose
One of the solutions simply parses the input string on each round (not
too costly, since we have to do < 100 rounds), the other one is
slightly faster, does the parsing of the movements before, but that
adds a second case machine, making the code a little too verbose.

Runtimes ~0.8 vs 0.5s.